Initial commit for new steam game tracker python api.
This is a python rewrite of the existing steam game tracker api (written in
nodejs).  This will live as an AWS lambda function.

Initially implemented the steam proxy api since it's very simple and
enables me to play and get the lambda working, but due to the way lambda
billing works, I may remove it and keep it hosted locally on my home server
where the tracker job lives.

This will eventually interface with some sort of AWS db (probably dynamo
due to pricing).